hide structure in detail namespace

+++ benchmark/Matlab/fecDecode.m
@@ -1,10 +1,10 @@
 function results = fecDecode(code, msg, llr, N, z)
-    decodedMsg = int8(code.decode(llr));
+    decodedMsg = code.decode(llr);
     elapsedTime = zeros(N,1);
     errorCount = zeros(N,1);
     for i = 1:N
         tic;
-        decodedMsg = int8(code.decode(llr));
+        decodedMsg = code.decode(llr);
         errorCount(i) = sum(sum(decodedMsg~=msg));
         elapsedTime(i) = toc;
     end
